requirements.txt
python-telegram-bot==21.6
anthropic>=0.86.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
apscheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.2
aiohttp==3.10.10
